chore: generate samples

+++ samples/client/petstore/csharp-netcore/OpenAPIClient-generichost-netcore-latest-oneOf/src/Org.OpenAPITools/Model/Banana.cs
@@ -118,7 +118,7 @@ namespace Org.OpenAPITools.Model
                     {
                         case "count":
                             if (utf8JsonReader.TokenType != JsonTokenType.Null)
-                                count = utf8JsonReader.GetDecimal();
+                                count = utf8JsonReader.GetInt32();
                             break;
                         default:
                             break;
